fix: merge gateway controller provider parameters into a single one
This is part of a global change in the project where all provider/controller/middleware
creators will only have one parameter: an "optional `options`" dictionary. In the case
of the `gatewayController` controller creator, the `middlewares` parameter is now part
of the `options` parameter.

BREAKING CHANGE: The `gatewayController` controller creator's `middleware` parameter no
longer exists; it's now part of the `options` parameter.