fix: make the forceHTTPS middleware parameter a dictionary of options.
This is part of a global change in the project where all provider/controller/middleware
creators will only have one paramter: an "optional `options`" dictionary. In the case
of the `forceHTTPS` middleware creator, the `ignoredRoutes` parameter is now part
of an `options` parameter.

BREAKING CHANGE: The `forceHTTPS` middleware creator's `ignoredRoutes` parameter has been
replaced with an object with a property `ignoredRoutes`.