docs: use a types file instead of a typedef
The problem with the typedef.js file is that the IDE will automatically load it for
all files located on the same level or below; if a file is on a higher level, it
won't get the the typedef types. Also, types from a typedef file can't be imported,
as there's no actual JS code on the file, the editor won't allow type imports, and
if you add some code, it invalids its "typedef status" and its types are not longer
globals.
The solution was to create a types.js, add an empty exports and import the types
from there; and for the couple of types that should be global, import them from
a "small typedef".